create b2b & b2c components
