+++ src/components/src/B2bHomePage/README.md
@@ -1,0 +1,21 @@
+# B2BHomePage
+
+#### Description
+
+Main static marketing content on home page.
+
+#### Usage
+
+```js
+import { B2BHomePage } from '@elasticpath/store-components';
+```
+
+#### Example
+
+```js
+<B2BHomePage />
+```
+
+#### Properties
+
+<!-- PROPS -->
