+++ src/components/src/B2cHomePage/README.md
@@ -1,0 +1,21 @@
+# B2CHomePage
+
+#### Description
+
+Main static marketing content on home page.
+
+#### Usage
+
+```js
+import { B2CHomePage } from '@elasticpath/store-components';
+```
+
+#### Example
+
+```js
+<B2CHomePage />
+```
+
+#### Properties
+
+<!-- PROPS -->
