+++ src/components/src/B2cHomePage/b2c.home.page.less
@@ -1,0 +1,323 @@
+/**
+ * Copyright © 2019 Elastic Path Software Inc. All rights reserved.
+ *
+ * This is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this license. If not, see
+ *
+ *     https://www.gnu.org/licenses/
+ *
+ *
+ */
+
+@import '../../../theme/common.less';
+
+.home-page-component {
+  height: 100%;
+
+  .section, .section-parallax {
+    max-width: 100%;
+    padding-right: 0px;
+    padding-left: 0px;
+    text-align: center;
+    display: table;
+    min-height: 260px;
+
+    .line {
+      display: block;
+      padding: 0 10px;
+    }
+
+    .line-1 {
+      font-size: 18px;
+      color: @mainColor;
+    }
+
+    .line-2 {
+      font-size: 12px;
+      line-height: 1.5;
+      color: #070707;
+      margin-top: 15px;
+    }
+
+    .line-3 {
+      padding: 20px;
+    }
+
+    .line-white {
+      font-size: 14px;
+      color: white;
+      letter-spacing: .4px;
+    }
+
+    .line-link {
+      color: @firstComplimentColor;
+      font-size: 12px;
+      margin-top: 20px;
+      position: relative;
+      display: inline-block;
+
+      &:after {
+        @size: 8px;
+        @border: 1.5px;
+
+        content: '';
+        display: block;
+        border-top: @border solid @firstComplimentColor;
+        border-right: @border solid @firstComplimentColor;
+        height: @size;
+        width: @size;
+        position: absolute;
+        transform: rotate(45deg) translate(0, -50%);
+        top: 45%;
+        right: 0;
+      }
+    }
+
+    .sub-section {
+      vertical-align: middle;
+      display: table-cell;
+
+      .text-block {
+        width: 320px;
+        display: inline-block;
+        text-align: left;
+        vertical-align: inherit;
+        padding-left: 20px;
+      }
+    }
+  }
+
+  .small-image {
+    width: 161px;
+    height: 161px;
+    margin-left: 30px;
+  }
+
+  .section-1 {
+    min-height: 0px;
+
+    .cover {
+      width: 100%;
+    }
+  }
+
+  .section-2 {
+    background-color: #f7f7f7;
+
+    .line-1 {
+      font-size: 24px;
+      font-weight: bold;
+      letter-spacing: -0.6px;
+      text-align: center;
+      color: #4a4a4a;
+    }
+
+    .line-2 {
+      font-size: 12px;
+      letter-spacing: -0.3px;
+      color: #4a4a4a;
+      margin-top: 20px
+    }
+
+    .btn-accessories {
+      width: 300px;
+      height: 40px;
+      background-color: @firstComplimentColor;
+      font-size: 12px;
+      font-weight: bold;
+      color: @mainBackgroundColor;
+      border: 0;
+      margin-top: 40px;
+      margin-bottom: 20px;
+      display: inline-block;
+
+      span {
+        font-size: 12px;
+        font-weight: bold;
+        color: @mainBackgroundColor;
+        vertical-align: -webkit-baseline-middle;
+        vertical-align: -moz-middle-with-baseline;
+      }
+
+      &:hover {
+        text-decoration: none;
+        cursor: pointer;
+      }
+    }
+  }
+
+  .section-parallax {
+    display: block;
+    overflow: hidden;
+    position: relative;
+    background-color: @mainColor;
+    height: 35vw;
+    max-height: 700px;
+
+    .parallax-image {
+      width: 100%;
+    }
+
+    .sub-section {
+      position: absolute;
+
+      .text-block {
+        padding-left: 0;
+      }
+    }
+  }
+
+  .section-parallax-1 {
+    .sub-section {
+      left: 18%;
+      top: 50%;
+      transform: translate(0, -50%);
+    }
+  }
+
+  .section-parallax-2 {
+    @img-container-width: 105vh;
+
+    .parallax-image-container {
+      width: @img-container-width;
+    }
+
+    .sub-section {
+      left: @img-container-width;
+      top: 50%;
+      transform: translate(0, -50%);
+    }
+  }
+
+  .section-4 {
+    background-color: #f7f7f7;
+  }
+
+  @media (max-width: @tabletWidth - 1px) {
+    .section-parallax {
+      height: 400px;
+    }
+
+    .section-parallax-1 {
+      .sub-section {
+        left: 8%;
+      }
+    }
+
+    .section-parallax-2 {
+      @img-container-width: 62%;
+
+      .parallax-image-container {
+        width: @img-container-width;
+
+        .parallax-image {
+          height: 500px;
+          object-position: right;
+          object-fit: cover;
+        }
+      }
+
+      .sub-section {
+        left: @img-container-width - 5%;
+        top: 65%;
+      }
+    }
+  }
+
+  @media (min-width: @mobileWidth) {
+    .small-image.image-1 {
+      display: none;
+    }
+  }
+
+  @media (max-width: @mobileWidth - 1px) {
+    .section {
+      min-height: 0px;
+
+      &.section-3, &.section-4, &.section-5 {
+        columns: 2;
+
+        .sub-section,
+        .text-block {
+          .small-image {
+            display: block;
+            margin-left:auto;
+            margin-right:auto;
+
+            &.image-2 {
+              display: none;
+            }
+          }
+        }
+      }
+
+      .small-image {
+        margin-top: 20px;
+      }
+
+      .line {
+        text-align: center;
+      }
+
+      .line-1 {
+        margin-top: 20px;
+      }
+
+      .line-2 {
+        margin-bottom: 20px;
+      }
+
+      .btn-accessories {
+        margin-top: 20px;
+      }
+    }
+
+    .section-parallax {
+      height: 60vh;
+
+      .parallax-image {
+        height: 70vh;
+        object-fit: cover;
+      }
+
+      .sub-section {
+        left: 0;
+        top: unset;
+        bottom: 10%;
+        transform: translate(0, 0);
+
+        .text-block {
+          width: 100%;
+        }
+      }
+    }
+
+    .section-parallax-1 {
+      .parallax-image {
+        height: 70vh;
+        object-position: 70%;
+      }
+    }
+
+    .section-parallax-2 {
+      .parallax-image-container {
+        width: 100%;
+
+        .parallax-image {
+          height: 55vh;
+          object-position: right;
+        }
+      }
+    }
+  }
+}
