+++ src/containers/HomePage.tsx
@@ -21,182 +21,23 @@
 
 import React from 'react';
 import { withRouter } from 'react-router';
-import intl from 'react-intl-universal';
-import { ParallaxProvider, Parallax } from 'react-scroll-parallax';
-import { ComplianceSupportModal, Carousel, IndiRecommendationsDisplayMain } from '../components/src/index';
+import {
+  B2BHomePage,
+  B2CHomePage,
+} from '../components/src/index';
 import Config from '../ep.config.json';
-
-import homeEspot2 from '../images/site-images/brake-icon.jpg';
-import homeEspot3 from '../images/site-images/charging-icon.jpg';
-import homeEspot4 from '../images/site-images/warranty-graphic.jpg';
-import homeEspotParallax1 from '../images/site-images/car-inside.jpg';
-import homeEspotParallax2 from '../images/site-images/car-dashboard.jpg';
 
 import './HomePage.less';
 
-const homeEspot2FileName = 'brake-icon.jpg';
-const homeEspot3FileName = 'charging-icon.jpg';
-const homeEspot4FileName = 'warranty-graphic.jpg';
-const homeEspotParallax1FileName = 'car-inside.jpg';
-const homeEspotParallax2FileName = 'car-dashboard.jpg';
-
-const showCompliance = Config.Compliance.enable;
-const isComplianceAccepted = localStorage.getItem(`${Config.cortexApi.scope}_Compliance_Accept`);
-const isComplianceDeclined = localStorage.getItem(`${Config.cortexApi.scope}_Compliance_Decline`);
-
 const HomePage: React.FunctionComponent = () => {
-  function handleAcceptDataPolicy() {
-    window.location.reload();
-  }
-  // Set the language-specific configuration for indi integration
-  Config.indi.brandAmbassador.title = intl.get('indi-brand-ambassador-title');
-  Config.indi.brandAmbassador.description = intl.get('indi-brand-ambassador-description');
-  Config.indi.brandAmbassador.submit_button_text = intl.get('indi-brand-ambassador-submit-button-text');
-
   if (!Config.b2b.enable) {
     return (
-      <ParallaxProvider>
-        <div className="home-page-component" data-region="viewPortRegion">
-          <Carousel />
-          <IndiRecommendationsDisplayMain render={['carousel', 'brand']} configuration={Config.indi} />
-          {/* eslint-disable-next-line no-return-assign */}
-          <div className="section section-3 container">
-            <div className="sub-section">
-              <img className="small-image" alt="home-espot-2" src={Config.siteImagesUrl.replace('%fileName%', homeEspot2FileName)} onError={(e: any) => { e.target.src = homeEspot3; }} />
-              <div className="text-block">
-                <span className="line line-1">
-                  {intl.get('home-sub-espot-container2-first-line')}
-                </span>
-                <span className="line line-2">
-                  {intl.get('home-sub-espot-container2-second-line')}
-                </span>
-              </div>
-            </div>
-          </div>
-          <div className="section-parallax section-parallax-1 container" data-region="homeMainContentRegion">
-            <Parallax y={[-50, 10]} tagOuter="figure">
-              <div className="parallax-image-container">
-                <img className="parallax-image" alt="home-espot-1" src={Config.siteImagesUrl.replace('%fileName%', homeEspotParallax1FileName)} onError={(e: any) => { e.target.src = homeEspot3; }} />
-              </div>
-            </Parallax>
-            <div className="sub-section">
-              <div className="text-block">
-                <span className="line line-white">
-                  {intl.get('home-sub-espot-container2-second-line')}
-                </span>
-                <span className="line line-link">
-                Learn more
-                </span>
-              </div>
-            </div>
-          </div>
-          <div className="section section-4 container">
-            <div className="sub-section">
-              <img className="small-image image-1" alt="home-espot-3" src={Config.siteImagesUrl.replace('%fileName%', homeEspot3FileName)} onError={(e: any) => { e.target.src = homeEspot3; }} />
-              <div className="text-block">
-                <span className="line line-1">
-                  {intl.get('home-sub-espot-container3-first-line')}
-                </span>
-                <span className="line line-2">
-                  {intl.get('home-sub-espot-container3-second-line')}
-                </span>
-              </div>
-              <img className="small-image image-2" alt="home-espot-3" src={Config.siteImagesUrl.replace('%fileName%', homeEspot3FileName)} onError={(e: any) => { e.target.src = homeEspot3; }} />
-            </div>
-          </div>
-          <div className="section section-5 container">
-            <div className="sub-section">
-              <img className="small-image" alt="home-espot-4" src={Config.siteImagesUrl.replace('%fileName%', homeEspot4FileName)} onError={(e: any) => { e.target.src = homeEspot4; }} />
-              <div className="text-block">
-                <span className="line line-1">
-                  {intl.get('home-sub-espot-container4-first-line')}
-                </span>
-                <span className="line line-2">
-                  {intl.get('home-sub-espot-container4-second-line')}
-                </span>
-              </div>
-            </div>
-          </div>
-          <div className="section-parallax section-parallax-2 container" data-region="homeMainContentRegion">
-            <Parallax y={[10, -30]} tagOuter="figure">
-              <div className="parallax-image-container">
-                <img className="parallax-image" alt="home-espot-1" src={Config.siteImagesUrl.replace('%fileName%', homeEspotParallax2FileName)} onError={(e: any) => { e.target.src = homeEspotParallax2; }} />
-              </div>
-            </Parallax>
-            <div className="sub-section">
-              <div className="text-block">
-                <span className="line line-white">
-                  {intl.get('home-sub-espot-container2-second-line')}
-                </span>
-                <span className="line line-link">
-                Learn more
-                </span>
-              </div>
-            </div>
-          </div>
-          {showCompliance && !isComplianceAccepted && !isComplianceDeclined && <ComplianceSupportModal onAcceptDataPolicy={handleAcceptDataPolicy} />}
-        </div>
-      </ParallaxProvider>
+      <B2CHomePage />
     );
   }
 
   return (
-    <div className="home-page-b2b">
-      <section className="main-banner">
-        <div className="main-banner-title-wrap">
-          <div className="container">
-            <h2 className="goods-heading">{intl.get('main-banner-heading')}</h2>
-            <div className="main-banner-txt">
-              <p className="goods-description">
-                {intl.get('main-banner-txt')}
-              </p>
-              <div className="btn-wrap">
-                <button type="button" className="ep-btn primary learn-more-btn">{intl.get('learn-more')}</button>
-              </div>
-            </div>
-          </div>
-        </div>
-      </section>
-
-      <div className="container">
-        <div className="main-goods">
-          <ul className="main-goods__grid">
-            <li className="main-goods__cell">
-              <h5 className="goods-title-small">{intl.get('products')}</h5>
-              <h3 className="goods-title">{intl.get('product1-heading')}</h3>
-              <p className="goods-description">{intl.get('product1-description')}</p>
-            </li>
-            <li className="main-goods__cell">
-              <h5 className="goods-title-small">{intl.get('products')}</h5>
-              <h3 className="goods-title">{intl.get('product2-heading')}</h3>
-              <p className="goods-description">{intl.get('product2-description')}</p>
-            </li>
-            <li className="main-goods__cell">
-              <h5 className="goods-title-small">{intl.get('products')}</h5>
-              <h3 className="goods-title">{intl.get('product3-heading')}</h3>
-              <p className="goods-description">
-                {intl.get('product3-description')}
-              </p>
-            </li>
-          </ul>
-
-          <div className="main-goods__cell main-goods__block">
-            <h5 className="goods-title-small">{intl.get('services')}</h5>
-            <h3 className="goods-title">{intl.get('service-heading')}</h3>
-            <p className="goods-description">
-              {intl.get('service-description')}
-            </p>
-          </div>
-
-          <div className="main-goods__cell main-goods__block bottom-block">
-            <h2 className="goods-heading">{intl.get('about-brand-heading')}</h2>
-            <p className="goods-description">
-              {intl.get('about-brand-description')}
-            </p>
-          </div>
-        </div>
-      </div>
-    </div>
+    <B2BHomePage />
   );
 };
